# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/proposals": {
      "get": {
        "description": "All proposals with client name, source (meeting/detection), totals, state, policy verdict and Xero link. Filter with ?state=.",
        "operationId": "ProposalsController_list",
        "parameters": [
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "DRAFT",
                "IN_REVIEW",
                "APPROVED",
                "SENT",
                "REJECTED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProposalSummaryDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List invoice proposals",
        "tags": [
          "proposals"
        ]
      }
    },
    "/proposals/{id}": {
      "get": {
        "description": "Full lines with provenance chips, policy reasons, the source evidence chain (meeting/transcript/detection/contract) and live Xero status.",
        "operationId": "ProposalsController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalDetailDto"
                }
              }
            }
          }
        },
        "summary": "Get a proposal in full",
        "tags": [
          "proposals"
        ]
      }
    },
    "/proposals/{id}/approve": {
      "post": {
        "description": "Writes an AUTHORISED ACCREC invoice (contact + invoice + decision note + evidence), sets state SENT, resolves the REVIEW_INVOICE task and marks the source meeting SENT. On Xero failure the proposal stays IN_REVIEW and the error is returned in the body (never a 500).",
        "operationId": "ProposalsController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalActionResultDto"
                }
              }
            }
          }
        },
        "summary": "Approve and send the proposal to Xero",
        "tags": [
          "proposals"
        ]
      }
    },
    "/proposals/{id}/save-draft": {
      "post": {
        "description": "Writes a DRAFT (unauthorised) ACCREC invoice to Xero and records its id + deep link, but keeps the proposal IN_REVIEW awaiting approval. On Xero failure the error is returned in the body (never a 500).",
        "operationId": "ProposalsController_saveDraft",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalActionResultDto"
                }
              }
            }
          }
        },
        "summary": "Create a DRAFT invoice in Xero",
        "tags": [
          "proposals"
        ]
      }
    },
    "/proposals/{id}/reject": {
      "post": {
        "description": "Sets state REJECTED, resolves the REVIEW_INVOICE task and records the reason on the audit trail.",
        "operationId": "ProposalsController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectProposalDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalActionResultDto"
                }
              }
            }
          }
        },
        "summary": "Reject the proposal",
        "tags": [
          "proposals"
        ]
      }
    },
    "/proposals/{id}/lines": {
      "patch": {
        "description": "Replaces the line set before approval and recomputes line amounts + subtotal, tax and total.",
        "operationId": "ProposalsController_patchLines",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchLinesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalActionResultDto"
                }
              }
            }
          }
        },
        "summary": "Edit the proposal lines",
        "tags": [
          "proposals"
        ]
      }
    },
...
  },
  "info": {
    "title": "Robyn API",
    "description": "Robyn turns a freelancer's calendar, contracts and transcripts into invoices in Xero. Every agent behaviour is a state transition that writes to Xero or raises a Task. The LLM parses and proposes at the edges; it never decides and never sends.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "PolicyResultDto": {
        "type": "object",
        "properties": {
          "auto": {
            "type": "boolean",
            "description": "Whether Robyn may send without asking."
          },
          "reasons": {
            "description": "Every reason the gate passed/failed.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "auto",
          "reasons"
        ]
      },
      "ProvenanceKind": {
        "type": "string",
        "enum": [
          "CONTRACT_CLAUSE",
          "TRANSCRIPT_QUOTE",
          "CALENDAR_BLOCK",
          "LEDGER"
        ]
      },
      "LineProvenanceDto": {
        "type": "object",
        "properties": {
          "kind": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ProvenanceKind"
              }
            ]
          },
          "label": {
            "type": "string",
            "example": "Clause 3.1"
          },
          "detail": {
            "type": "string",
            "example": "The Consultant shall be paid £150 per hour."
          },
          "source_ref": {
            "type": "string",
            "description": "Source id (meeting/contract/quote)."
          }
        },
        "required": [
          "kind",
          "label",
          "detail"
        ]
      },
      "ProposalLineDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "quantity": {
            "type": "number",
            "example": 1.5
          },
          "unit_amount": {
            "type": "number",
            "example": 150
          },
          "line_amount": {
            "type": "number",
            "example": 225
          },
          "account_code": {
            "type": "string",
            "example": "200"
          },
          "tax_type": {
            "type": "string",
            "example": "OUTPUT2"
          },
          "provenance": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LineProvenanceDto"
            }
          }
        },
        "required": [
          "description",
          "quantity",
          "unit_amount",
          "line_amount",
          "provenance"
        ]
      },
...
      "ProposalSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "clientId": {
            "type": "string"
          },
          "clientName": {
            "type": "string",
            "description": "Resolved client display name"
          },
          "source": {
            "type": "string",
            "enum": [
              "meeting",
              "detection"
            ],
            "description": "What this proposal was built from"
          },
          "meetingId": {
            "type": "object",
            "nullable": true,
            "description": "Source meeting id, if any"
          },
          "detectionId": {
            "type": "object",
            "nullable": true,
            "description": "Source detection id, if any"
          },
          "currency": {
            "type": "string",
            "example": "GBP"
          },
          "subtotal": {
            "type": "number",
            "example": 225
          },
          "taxTotal": {
            "type": "number",
            "example": 45
          },
          "total": {
            "type": "number",
            "example": 270
          },
          "state": {
            "type": "string",
            "enum": [
              "DRAFT",
              "IN_REVIEW",
              "APPROVED",
              "SENT",
              "REJECTED"
            ]
          },
          "policyResult": {
            "$ref": "#/components/schemas/PolicyResultDto"
          },
          "lineCount": {
            "type": "number",
            "description": "Number of invoice lines"
          },
          "autoSent": {
            "type": "boolean",
            "description": "Whether Robyn auto-sent this (autonomy ON)"
          },
          "reference": {
            "type": "object",
            "nullable": true,
            "description": "Idempotency reference used in Xero"
          },
          "xeroInvoiceId": {
            "type": "object",
            "nullable": true
          },
          "xeroInvoiceNumber": {
            "type": "object",
            "nullable": true
          },
          "xeroDeepLink": {
            "type": "object",
            "nullable": true,
            "description": "Deep link into the Xero UI"
          },
          "createdAt": {
            "type": "string",
            "description": "ISO created timestamp"
          }
        },
        "required": [
          "id",
          "clientId",
          "clientName",
          "source",
          "currency",
          "subtotal",
          "taxTotal",
          "total",
          "state",
          "policyResult",
          "lineCount",
          "autoSent",
          "createdAt"
        ]
      },
      "ProposalEvidenceMeetingDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "start": {
            "type": "string",
            "description": "ISO start"
          },
          "end": {
            "type": "string",
            "description": "ISO end"
          },
          "durationHours": {
            "type": "number"
          },
          "state": {
            "type": "string"
          },
          "attendees": {
            "description": "Attendee emails",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "title",
          "start",
          "end",
          "durationHours",
          "state",
          "attendees"
        ]
      },
      "ProposalEvidenceTranscriptDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "summary": {
            "type": "object",
            "nullable": true
          },
          "scopeQuotes": {
            "description": "Verbatim billable-scope quotes",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "excerpt": {
            "type": "string",
            "description": "Short raw excerpt for the drill-down"
          }
        },
        "required": [
          "id",
          "scopeQuotes",
          "excerpt"
        ]
      },
      "ProposalEvidenceDetectionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "valueGbp": {
            "type": "number"
          },
          "summary": {
            "type": "string",
            "description": "Detector evidence summary"
          }
        },
        "required": [
          "id",
          "type",
          "valueGbp",
          "summary"
        ]
      },
      "ProposalEvidenceContractDto": {
        "type": "object",
        "properties": {
          "contractId": {
            "type": "string"
          },
          "clauseLabel": {
            "type": "object",
            "nullable": true
          },
          "clauseText": {
            "type": "object",
            "nullable": true,
            "description": "Verbatim clause text"
          }
        },
        "required": [
          "contractId"
        ]
      },
      "ProposalEvidenceDto": {
        "type": "object",
        "properties": {
          "meeting": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ProposalEvidenceMeetingDto"
              }
            ]
          },
          "transcript": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ProposalEvidenceTranscriptDto"
              }
            ]
          },
          "detection": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ProposalEvidenceDetectionDto"
              }
            ]
          },
          "contract": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ProposalEvidenceContractDto"
              }
            ]
          }
        }
      },
      "ProposalDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "clientId": {
            "type": "string"
          },
          "clientName": {
            "type": "string",
            "description": "Resolved client display name"
          },
          "source": {
            "type": "string",
            "enum": [
              "meeting",
              "detection"
            ],
            "description": "What this proposal was built from"
          },
          "meetingId": {
            "type": "object",
            "nullable": true,
            "description": "Source meeting id, if any"
          },
          "detectionId": {
            "type": "object",
            "nullable": true,
            "description": "Source detection id, if any"
          },
          "currency": {
            "type": "string",
            "example": "GBP"
          },
          "subtotal": {
            "type": "number",
            "example": 225
          },
          "taxTotal": {
            "type": "number",
            "example": 45
          },
          "total": {
            "type": "number",
            "example": 270
          },
          "state": {
            "type": "string",
            "enum": [
              "DRAFT",
              "IN_REVIEW",
              "APPROVED",
              "SENT",
              "REJECTED"
            ]
          },
          "policyResult": {
            "$ref": "#/components/schemas/PolicyResultDto"
          },
          "lineCount": {
            "type": "number",
            "description": "Number of invoice lines"
          },
          "autoSent": {
            "type": "boolean",
            "description": "Whether Robyn auto-sent this (autonomy ON)"
          },
          "reference": {
            "type": "object",
            "nullable": true,
            "description": "Idempotency reference used in Xero"
          },
          "xeroInvoiceId": {
            "type": "object",
            "nullable": true
          },
          "xeroInvoiceNumber": {
            "type": "object",
            "nullable": true
          },
          "xeroDeepLink": {
            "type": "object",
            "nullable": true,
            "description": "Deep link into the Xero UI"
          },
          "createdAt": {
            "type": "string",
            "description": "ISO created timestamp"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProposalLineDto"
            }
          },
          "evidence": {
            "$ref": "#/components/schemas/ProposalEvidenceDto"
          },
          "xeroStatus": {
            "type": "object",
            "nullable": true,
            "description": "Live Xero status (DRAFT/AUTHORISED/...) if written"
          },
          "xeroError": {
            "type": "object",
            "nullable": true,
            "description": "Last Xero write error, if the write failed"
          }
        },
        "required": [
          "id",
          "clientId",
          "clientName",
          "source",
          "currency",
          "subtotal",
          "taxTotal",
          "total",
          "state",
          "policyResult",
          "lineCount",
          "autoSent",
          "createdAt",
          "lines",
          "evidence"
        ]
      },
      "ProposalActionResultDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "enum": [
              "DRAFT",
              "IN_REVIEW",
              "APPROVED",
              "SENT",
              "REJECTED"
            ]
          },
          "xeroWritten": {
            "type": "boolean",
            "description": "Did the Xero write succeed?"
          },
          "xeroInvoiceId": {
            "type": "object",
            "nullable": true
          },
          "xeroInvoiceNumber": {
            "type": "object",
            "nullable": true
          },
          "xeroDeepLink": {
            "type": "object",
            "nullable": true
          },
          "xeroError": {
            "type": "object",
            "nullable": true,
            "description": "Xero error message when the write failed"
          },
          "subtotal": {
            "type": "number",
            "example": 225
          },
          "taxTotal": {
            "type": "number",
            "example": 45
          },
          "total": {
            "type": "number",
            "example": 270
          },
          "message": {
            "type": "string",
            "description": "Human sentence describing what happened"
          }
        },
        "required": [
          "id",
          "state",
          "xeroWritten",
          "subtotal",
          "taxTotal",
          "total",
          "message"
        ]
      },
      "RejectProposalDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "Client already paid this in cash — do not invoice.",
            "description": "Why the proposal was rejected (kept on the audit trail)"
          }
        },
        "required": [
          "reason"
        ]
      },
      "PatchLinesDto": {
        "type": "object",
        "properties": {
          "lines": {
            "description": "The full edited line set",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProposalLineDto"
            }
          }
        },
        "required": [
          "lines"
        ]
      },
...
    }
  }
}